backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic
python-jose
passlib[bcrypt]==1.7.4
bcrypt==3.2.2

python-dotenv
email-validator
